use details module in file itself
